+++ Sources/KeyboardKit/_Deprecated/AutocompleteSuggestionProvider+Deprecated.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "AutocompleteProvider")
+public typealias AutocompleteSuggestionProvider = AutocompleteProvider