+++ Sources/KeyboardKit/_Deprecated/KeyboardInputViewController+Deprecated.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public extension KeyboardInputViewController {
+    
+    @available(*, deprecated, renamed: "autocompleteProvider")
+    var autocompleteSuggestionProvider: AutocompleteProvider {
+        get { autocompleteProvider }
+        set { autocompleteProvider = newValue }
+    }
+}